多个Visindigo功能更新 1. 初步完成命令处理功能，在虚拟终端初步增加相应的命令回溯等功能，目前已知有bug需要处理 2. 为Exception、Logger增加一些API，并提供了崩溃日志功能 3. 强化ThemeManager，增加调色板切换时的动画，并在Plugin增加了相应注册资源的API 4. 在VIMacro增加了一些标记宏 5. 取消了GeneralConfig的数据类内存池

+++ src/Visindigo/General/private/LoggerMsgHandler_impl.hpp
@@ -12,7 +12,7 @@ namespace Visindigo::General {
 		QString temp = "QMap{";
 		for (QString key : pointer_map) {
 			void* ptr = &pointer_map[key];
-			temp += " \"" % key % "\": [Object at " % QString::number(ptr, 16) % "],";
+			temp += " \"" % key % "\": [Object at " % QString::number((quint64)ptr, 16) % "],";
 		}
 		temp.removeLast();
 		temp += "}";
